zbackend/requirements.txt
Django==5.0.4
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
django-allauth==0.61.1
drf-yasg==1.21.7
dj-rest-auth==6.0.0
sqlparse==*
psycopg2-binary==*
python-dotenv==1.0.1
uwsgi>=2.0.19<2.1
